pyproject.toml
[project]
name = "alphaswarm"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex Mark", email = "alexandermarx@chainml.net"},
    {name = "Arnaud Flament", email = "arnaudflament@chainml.net"},
    {name = "Daniel Kur", email = "danielkur@chainml.net"},
    {name = "David Mueller", email = "davidmueller@chainml.net"},
    {name = "Dmytro Nikolaiev", email = "dmytronikolaiev@chainml.net"},
    {name = "Ethan Jackson", email = "ethanjackson@chainml.net"},
    {name = "Guillaume Koch", email = "guillaumekoch@chainml.net"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "pyyaml (>=6.0.0,<7.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
]

[tool.poetry.group.ai.dependencies]
transformers = "4.47.1"  # fixed smolagents dependency
litellm = "1.55.10"  # newer versions via smolagents will not work on windows
instructor = "1.7.2"
smolagents = "1.2.2"

[tool.poetry.group.clients.dependencies]
python-telegram-bot = "21.10"

[tool.poetry.group.web3.dependencies]
web3 = "6.14.0"
solana = "0.36.0"
web3-ethereum-defi = "0.28"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
types-pyyaml = "^6.0.12.20241230"
ruff = "^0.9.5"
mypy = "^1.15.0"
black = "^24.10.0"
isort = "^5.13.2"
pytest = "^8.3.4"
types-requests = ">=2.31.0.6"
python-dotenv = "^1.0.1"
eth-typing = "3.5.2"
pytest-html = "^4.1.1"
pytest-cov = "^6.0.0"

[tool.poetry]
include = [
    { path = "config/default.yaml", format = "wheel" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

#[tool.pytest]
#pythonpath = "."

[tool.pytest.ini_options]
log_cli = "True"
log_cli_level = "DEBUG"
log_cli_format = "[%(asctime)s %(levelname)s %(threadName)s %(name)s:%(funcName)s:%(lineno)s] %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S%z"

[tool.black]
line-length = 120

[tool.mypy]
exclude = "(?x)(venv|docs|tmp)"

explicit_package_bases = "true"
no_implicit_optional = true
disallow_untyped_defs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["smolagents.*", "theoriq.*", "eth_defi.*"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
exclude = ["__init__.py", "docs", "tmp", "venv"]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
py_version = 311
src_paths = ["src"]
skip = ["__init__.py"]
skip_glob = ["**/.venv/*", "**/venv/*", "**/docs/*", "**/stubs/*", "**/tests/*"]
